requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
asgiref==3.8.1
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
Brotli==1.1.0
build==1.2.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
coloredlogs==15.0.1
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
contextlib2==21.6.0
crewai==0.28.8
crewai-tools==0.1.7
cryptography==42.0.5
dataclasses-json==0.6.4
datasets==2.19.0
debugpy @ file:///work/perseverance-python-buildout/croot/debugpy_1698884710808/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
deepdiff==7.0.1
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.8
distro==1.9.0
docstring-parser==0.15
embedchain==0.1.102
emoji==2.11.1
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1704921103267/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
fastapi==0.110.2
filelock==3.13.4
filetype==1.2.0
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-aiplatform==1.48.0
google-cloud-bigquery==3.21.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
gptcache==0.1.43
greenlet==3.0.3
groq==0.5.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.2
grpcio-status==1.62.2
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.5
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.7
importlib-metadata==7.0.0
importlib_resources==6.4.0
iniconfig==2.0.0
instructor==0.5.2
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1708996548741/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1709559745751/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
joblib==1.4.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1710255804825/work
jupyter_core @ file:///work/perseverance-python-buildout/croot/jupyter_core_1701731747496/work
kubernetes==29.0.0
lancedb==0.5.7
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.46
langchain-google-vertexai==1.0.2
langchain-groq==0.1.3
langchain-openai==0.0.5
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langdetect==1.0.9
langsmith==0.1.50
lxml==5.2.1
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
mdurl==0.1.2
mediawikiapi==1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mutagen==1.47.0
mypy-extensions==1.0.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
nltk==3.8.1
nodeenv==1.8.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.3
openai==1.23.3
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
ordered-set==4.1.0
orjson==3.10.1
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.2
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1713912794367/work
playwright==1.43.0
pluggy==1.5.0
posthog==3.5.0
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1702399386289/work
proto-plus==1.23.0
protobuf==4.25.3
psutil @ file:///work/perseverance-python-buildout/croot/psutil_1698863411559/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pulsar-client==3.5.0
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
py==1.11.0
pyarrow==16.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
pyee==11.1.0
PyGithub==1.59.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1700607939962/work
PyJWT==2.8.0
pylance==0.9.18
PyNaCl==1.5.0
pypdf==4.2.0
PyPika==0.48.9
pyproject_hooks==1.0.0
pyright==1.1.360
pysbd==0.3.4
PySocks==1.7.1
pytest==8.1.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
python-dotenv==1.0.0
python-iso639==2024.2.7
python-magic==0.4.27
pytube==15.0.0
pytz==2024.1
PyYAML==6.0.1
pyzmq @ file:///croot/pyzmq_1705605076900/work
ragas==0.1.7
rapidfuzz==3.8.1
ratelimiter==1.2.0.post0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
retry==0.9.2
rich==13.7.1
rsa==4.9
safetensors==0.4.3
schema==0.7.5
selenium==4.20.0
semver==3.0.2
setuptools==68.2.2
shapely==2.0.4
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.29
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
starlette==0.37.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.5.2
tokenizers==0.19.1
tornado @ file:///work/perseverance-python-buildout/croot/tornado_1698866362018/work
tqdm==4.66.2
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
transformers==4.40.2
trio==0.25.0
trio-websocket==0.11.1
typer==0.9.4
types-protobuf==4.25.0.20240417
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1712329955671/work
tzdata==2024.1
unstructured==0.11.8
unstructured-client==0.22.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
vertexai==1.48.0
watchfiles==0.21.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
websocket-client==1.8.0
websockets==12.0
wheel==0.41.2
wikibase-rest-api-client==0.1.3
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.4
youtube-transcript-api==0.6.2
yt-dlp==2023.12.30
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1695255097490/work
